Remove redundant SkillsView.cs and consolidate logic
- Deleted `SkillsView.cs` and its `.meta` file as it's been superseded by `PlayerSkillView.cs`.
- Moved the `SkillsPayload` struct into `PlayerSkillView.cs` to maintain functionality with `SkillsScreenController.cs`.
- Created Context-Dense Metadata Summaries in `Documentation/Script_Descriptions/` for the remaining skill scripts: `PlayerSkillView.cs`, `SkillsScreenController.cs`, `SkillMenuController.cs`, and `SkillDataSO.cs`.
- Updated `CHANGELOG.md` to reflect these changes.
