--- Toris/Assets/Scripts/UIToolkit/UI/UIViews/SkillsView.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8054b42684d4f5c47999a9406f831a2e
